# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Virtual General Contractor API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quotes:
    get:
      operationId: listQuotes
      tags: [quotes]
      summary: List user quotes
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Quote list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteList"
  /quotes/generate:
    post:
      operationId: generateQuote
      tags: [quotes]
      summary: Generate a renovation quote from conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateQuoteBody"
      responses:
        "201":
          description: Quote generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /quotes/{quoteId}:
    get:
      operationId: getQuote
      tags: [quotes]
      summary: Get quote details
      parameters:
        - name: quoteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quote details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteQuote
      tags: [quotes]
      summary: Delete a quote
      parameters:
        - name: quoteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /quotes/shared/{shareUuid}:
    get:
      operationId: getSharedQuote
      tags: [quotes]
      summary: Get a shared quote (public, no auth required)
      parameters:
        - name: shareUuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Shared quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SharedQuote"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /quotes/{quoteId}/share:
    post:
      operationId: toggleShareQuote
      tags: [quotes]
      summary: Enable or disable sharing for a quote
      parameters:
        - name: quoteId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleShareBody"
      responses:
        "200":
          description: Share status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareStatus"
...
  /cost-engine/regional-multiplier:
    get:
      operationId: getRegionalMultiplier
      tags: [costEngine]
      summary: Get regional cost multiplier for a zip code
      parameters:
        - name: zipCode
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Regional multiplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegionalMultiplier"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Property:
      type: object
      properties:
        id:
          type: integer
        zillowUrl:
          type: ["string", "null"]
        address:
          type: string
        zipCode:
          type: string
        sqft:
          type: integer
        bedrooms:
          type: integer
        bathrooms:
          type: number
        yearBuilt:
          type: ["integer", "null"]
        lotSize:
          type: ["number", "null"]
        listingPhotos:
          type: array
          items:
            type: string
        priceHistory:
          type: ["object", "null"]
        dataSource:
          type: string
          enum: [apify, rentcast, manual]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - address
        - zipCode
        - sqft
        - bedrooms
        - bathrooms
        - dataSource
        - createdAt
...
    GenerateQuoteBody:
      type: object
      properties:
        conversationId:
          type: integer
        qualityTier:
          type: string
          enum: [economy, mid_range, premium]
          default: mid_range
        title:
          type: string
      required:
        - conversationId
    QuoteLineItem:
      type: object
      properties:
        id:
          type: integer
        category:
          type: string
        description:
          type: string
        materialCost:
          type: number
        laborCost:
          type: number
        quantity:
          type: number
        unit:
          type: string
        qualityTier:
          type: string
          enum: [economy, mid_range, premium]
        subtotal:
          type: number
      required:
        - id
        - category
        - description
        - materialCost
        - laborCost
        - quantity
        - unit
        - qualityTier
        - subtotal
    Quote:
      type: object
      properties:
        id:
          type: integer
        shareUuid:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
          enum: [draft, final]
        totalEstimate:
          type: number
        qualityTier:
          type: string
          enum: [economy, mid_range, premium]
        aiModelUsed:
          type: string
        aiReasoning:
          type: string
        regionalMultiplier:
          type: number
        sharedUrlEnabled:
          type: boolean
        property:
          $ref: "#/components/schemas/Property"
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/QuoteLineItem"
        claudeReview:
          type: ["object", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - shareUuid
        - title
        - status
        - totalEstimate
        - qualityTier
        - aiModelUsed
        - sharedUrlEnabled
        - lineItems
        - createdAt
        - updatedAt
    QuoteList:
      type: object
      properties:
        quotes:
          type: array
          items:
            $ref: "#/components/schemas/QuoteSummary"
        total:
          type: integer
      required:
        - quotes
        - total
    QuoteSummary:
      type: object
      properties:
        id:
          type: integer
        shareUuid:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
        totalEstimate:
          type: number
        qualityTier:
          type: string
        propertyAddress:
          type: string
        sharedUrlEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - shareUuid
        - title
        - status
        - totalEstimate
        - qualityTier
        - propertyAddress
        - sharedUrlEnabled
        - createdAt
    SharedQuote:
      type: object
      properties:
        quote:
          $ref: "#/components/schemas/Quote"
        agentName:
          type: ["string", "null"]
        agentBrokerage:
          type: ["string", "null"]
        coBrandName:
          type: ["string", "null"]
      required:
        - quote
    ToggleShareBody:
      type: object
      properties:
        enabled:
          type: boolean
      required:
        - enabled
    ShareStatus:
      type: object
      properties:
        sharedUrlEnabled:
          type: boolean
        shareUrl:
          type: ["string", "null"]
      required:
        - sharedUrlEnabled
...
    RegionalMultiplier:
      type: object
      properties:
        zipPrefix:
          type: string
        metroArea:
          type: string
        adjustmentFactor:
          type: number
        source:
          type: string
        lastUpdated:
          type: string
          format: date-time
      required:
        - zipPrefix
        - metroArea
        - adjustmentFactor
        - source
        - lastUpdated