# lib/api-zod/src/generated/types/generateQuoteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Virtual General Contractor API
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateQuoteBodyQualityTier } from "./generateQuoteBodyQualityTier";

export interface GenerateQuoteBody {
  conversationId: number;
  qualityTier?: GenerateQuoteBodyQualityTier;
  title?: string;
}