# lib/api-zod/src/generated/types/quoteList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Virtual General Contractor API
 * OpenAPI spec version: 0.1.0
 */
import type { QuoteSummary } from "./quoteSummary";

export interface QuoteList {
  quotes: QuoteSummary[];
  total: number;
}